refacotor : 데이터 흐름에 따라 컴포넌트 수정
