+++ src/components/Letter/Post/PostLetterForm.tsx
@@ -4,6 +4,8 @@ import { SelectSlider } from '../SelectSlier/SelectSlider';
 import React from 'react';
 import { useToastStore } from '@/hooks/useToastStore';
 import { TextArea } from '@/components/Common/TextArea/TextArea';
+import { TopBar } from '@/components/Common/TopBar/TopBar';
+import { useNavigate } from 'react-router-dom';
 
 export const PostLetterForm = () => {
     const [title, setTitle] = useState('');
@@ -11,6 +13,8 @@ export const PostLetterForm = () => {
     const [letterContent, setLetterContent] = useState<string>('');
     const [font, setFont] = useState<string>('');
     const { addToast } = useToastStore();
+
+    const navigate = useNavigate();
 
     const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         const inputValue = e.target.value;
@@ -22,31 +26,42 @@ export const PostLetterForm = () => {
     };
 
     return (
-        <div className="min-h-screen z-10rounded-t-3xl bg-zinc-300">
-            <Margin top={10} />
-            <div className="relative flex flex-col justify-center w-9/12 m-auto py-14">
-                <input
-                    onChange={handleChange}
-                    value={title}
-                    type="text"
-                    placeholder="제목을 입력해주세요"
-                    className="z-10 w-full bg-transparent border-none focus:border-none focus:outline-none text-wrap"
-                />
-                <img src={'/public/to_line.f4c129e6.svg'} />
-
-                <div className="relative z-10">
-                    <TextArea
-                        value={letterContent}
-                        setValue={setLetterContent}
-                    />
-                </div>
-            </div>
-            <SelectSlider
-                font={font}
-                letter={letter}
-                setFont={setFont}
-                setLetter={setLetter}
+        <>
+            <TopBar
+                handleBackClick={() => {
+                    navigate(-1);
+                }}
+                handelSuccesClick={() => {
+                    navigate('/letter/select');
+                }}
             />
-        </div>
+            <div className="min-h-screen rounded-t-3xl bg-zinc-300">
+                <Margin top={20} />
+                <div className="relative flex flex-col justify-center w-9/12 m-auto py-14">
+                    <input
+                        onChange={handleChange}
+                        value={title}
+                        type="text"
+                        placeholder="제목을 입력해주세요"
+                        className="z-10 w-full bg-transparent border-none focus:border-none focus:outline-none text-wrap"
+                    />
+                    <img src={'/public/to_line.f4c129e6.svg'} />
+
+                    <div className="relative z-10">
+                        <TextArea
+                            value={letterContent}
+                            setValue={setLetterContent}
+                            font={font}
+                        />
+                    </div>
+                </div>
+                <SelectSlider
+                    font={font}
+                    letter={letter}
+                    setFont={setFont}
+                    setLetter={setLetter}
+                />
+            </div>
+        </>
     );
 };
